DEU scrapers: add _shared.js, use in DE-BY scraper
First step for sharing code across DEU state
scrapers.
